move settings code into own module
